Refactor: Вынес логику модалки в хук
