+++ src/components/layouts/modal/modal.tsx
@@ -1,7 +1,6 @@
-import React, { PropsWithChildren, useEffect } from 'react'
+import { PropsWithChildren } from 'react'
 import ReactDOM from 'react-dom'
-import { useAppDispatch, useAppSelector } from './../../../redux/index'
-import { ModalPayload } from './../../../redux/reducers/modal-reducer/modal-reducer'
+import { useDialogHandling } from './hooks/use-dialog-handling'
 
 const portal = document.getElementById('portal')
 
@@ -9,33 +8,12 @@ interface ModalProps extends PropsWithChildren {
   isOpen?: boolean
 }
 
-function Modal({ isOpen, children }: ModalProps): JSX.Element {
-  const modalState = useAppSelector<ModalPayload>((state) => state.modalReducer)
-  const dispatch = useAppDispatch()
-  const modal = React.createRef<HTMLDialogElement>()
-
-  const onWindowEscKeydownHandler = (event: globalThis.KeyboardEvent): void => {
-    if (event.key === 'Esc' || event.key === 'Escape') {
-      dispatch({ type: 'CLOSE_MODAL' })
-    }
-  }
-
-  useEffect(() => {
-    window.addEventListener('keydown', onWindowEscKeydownHandler)
-    if (modalState.isOpen) {
-      modal.current?.showModal()
-    } else {
-      modal.current?.close()
-    }
-
-    return () => {
-      window.removeEventListener('keydown', onWindowEscKeydownHandler)
-    }
-  }, [modalState.isOpen])
+function Modal({ children }: ModalProps): JSX.Element {
+  const { modalRef } = useDialogHandling()
 
   if (portal !== null) {
     return ReactDOM.createPortal(
-      <dialog className='modal' ref={modal}>
+      <dialog className='modal' ref={modalRef}>
         <div className='modal__layout'>{children}</div>
       </dialog>,
       portal
